+++ common/src/main/java/cn/acitrus/common/entities/connector/connection/DefaultConnection.java
@@ -1,6 +1,7 @@
-package cn.acitrus.common.entities.config.connection;
+package cn.acitrus.common.entities.connector.connection;
 
 import cn.acitrus.common.enums.permissions.ConnectorType;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.AllArgsConstructor;
@@ -15,10 +16,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class RedisConnection implements Connection {
-    private String host;
-    private Integer port;
-    private String username;
-    private String password;
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class DefaultConnection implements Connection {
     private ConnectorType connectorType;
 }
