+++ common/src/main/java/cn/acitrus/common/entities/connector/connection/Neo4jConnection.java
@@ -1,4 +1,4 @@
-package cn.acitrus.common.entities.config.connection;
+package cn.acitrus.common.entities.connector.connection;
 
 import cn.acitrus.common.enums.permissions.ConnectorType;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
