--- transfer/src/main/java/cn/acitrus/transfer/projection/config/ConnectorEntityInfo.java
@@ -1,16 +1,0 @@
-package cn.acitrus.transfer.projection.config;
-
-import cn.acitrus.common.entities.config.connection.Connection;
-
-import java.util.UUID;
-
-/**
- * A Projection for the {@link ConnectorEntity} entity
- */
-public interface ConnectorEntityInfo {
-    UUID getId();
-
-    Integer getVersion();
-
-    Connection getConnection();
-}
